chore: improve Box exports
